inserted new SolrDefault layer and moved ServiceLocatorAwareInterface there
